Remove DOM update files from script.js
